Focusing bugs fixed; Auto compiling libraries added.
When selecting a calendar, or switching views, originally the focus was still set to the last thing clicked. This would require an extra click to activate any buttons in the calendar. Now when the calendar or the view is changed, blur() is ran. This fixes needing and extra click.

Paths in the tsconfig were pointed to the public_api of the libraries so that when any changes are saved in development, the app recompiles with out needing repackaging.